[INITIAL] first working implementation of ebtree
ebtree stands for Elastic Binary Tree. It's a new tree type
in which data is carried in nodes which can be split in two to
provide intermediate nodes for other values. It looks like a
radix tree, but is used like a binary tree (no malloc).

This first version does work but is not optimal WRT memory
usage, since both half nodes store the data and the bit value.
An optimal solution would only store either and provide a fast
solution to differentiate between an intermediate node and a
leaf node.