[OPTIM] prevent useless data copy during removal
When replacing a link node during a removal, it is not necessary
to recopy the value to the new node because it will necessarily
have been found blow the same link node, implying that it shares
the same higer bits. Also, it may be useful later to keep the
value.