[MAJOR] add support for unique keys
If an insert() is performed on a tree which has its right
branch set to (void *)1, then the node will not be inserted
if another node with the same key already exists. Instead, a
pointer to the already existing node is returned. This helps
a lot for counters.

The normal method to declare such a tree is to declare its
head with EB_ROOT_UNIQUE instead of EB_ROOT.