[MEDIUM] make ebpttree rely solely on eb32/eb64 trees
The functions are now defined inline and optimized at compile time
depending on the size of a pointer. This saves a lot of code in
programs using pointer trees simultaneously with integer trees.

The node types remain ebpt_nodes though, but there is no need to
have an ebpttree.c file anymore.