[MAJOR] switch the node's bit type from int to short int.
We're not interested in processing strings longer than 4 kB, so let's
reduce the bit type size to 16 bits in order to save some space to
introduce the prefix length into the node (16 bits too). This one
will be used to perform prefix-based longest-match, which is typically
used in network address lookups.

This change alone had significant impacts on the quality of the
resulting object code because gcc does not natively handle sub-int types,
and spends its time converting them to full 32-bit registers that it never
uses. So most of the change consisted in creating local 32-bit variables
to keep a 32-bit copy of the bit value so that gcc stops wasting CPU cycles
doing useless things.

Now the measured impact is around 0 to 2 cycles per lookup or insertion
(among 60 to 100), and the 32-bit code has even benefitted from the local
register assignment, by speeding up by about 10% the lookup operations.