[MEDIUM] ebmbtree: add prefix-based insertion/lookup
These new functions add the ability to store prefixes and find entries
by longest matches or exact prefix match. This will typically be used
to store network routing information. The walk, delete and duplicate
mechanisms are compatible with the ones used by the exact match functions.
The new features work by storing half-bits in the tree. A prefix is stored
as a "cover node" above the split point corresponding to its bit depth. It
means that -at least now- the exact-match insert/lookup functions are not
compatible with data stored in the prefix tree and conversely. However this
should not be an issue.

The new functions are :

   - ebmb_insert_prefix() : insert a prefix key of specified length in the
     tree. If a similar key/prefix entry already exists, either the tree
     supports duplicates and a dup is created, or the insertion is aborted
     and the first matching key is reported ;

   - ebmb_lookup_prefix() : looks for the exact key/prefix pair in the tree.
     The first exact match is returned, otherwise NULL is returned. The only
     likely use of this function is to look up for exact entries to be
     removed from the tree.

   - ebmb_lookup_longest() : looks for the first occurrence of the longest
     prefix which matches key. Null is returned if none is found.

Right now the feature has not been implemented for indirect memory blocks
(ebim_*), but it does not seem to be really useful since prefix will always
be of known length. Maybe it could make sense for space saving in complex
IPv6-based trees but that's not really sure.