[OPTIM] ebtree: add an flsnz8() function to speed up some block comparisons
The code emitted by gcc on 8-bit computations is even worse than what it
does on 32-bit. It does multiple needless conversions and tests. So add
generic and arch-specific flsnz8() functions to save gcc from having to
think. In the end, we save about 13% performance on the ebmb insertion
time.