[OPTIM] ebtree: improve equal_bits() by reducing register usage
By reworking equal_bits(), we can avoid converting bytes to bits in the
loop and save a few precious registers. On some architectures short in
registers such as x86, this avoids load/stores to/from memory for local
variables, and enables complex computations which are possible with
registers. This optimisation alone has made ebmb_insert() 16% faster,
and a few tens of bytes shorter.

The same principle should be applied to string_equal_bits().