[OPTIM] ebmbtree: don't compute bit position in lookup
It's a waste of time to compute a bit position during the tree descent
because we only use it for a comparison. By carefully doing a byte-to-byte
comparison and ensuring that we add no local variable to store values that
don't fit into registers, we can almost double the speed of the lookup
(measurements have shown CPU savings of 45%).