[MINOR] ebtree: add a check_bits() function to compare two blocks
This function compares two blocks on their first common bits. It is also
able to skip a known byte length, because it is intended to be used in
tree descent code where a number of full bytes has already been matched.
It has been verified to produce optimal code on gcc 2.95 to 4.2.