[RELEASE] version 6.0
This version adds support for prefix-based matching of memory blocks.
It provides a prefix insertion and longest match which are compatible
with the rest of the common features (walk, duplicates, delete, ...).
This is typically used for network address matching. The code is a bit
slower than the original memory block handling code, so they have not
been merged together into generic code. Still it's possible to perform
about 10 million networks lookups per second in a set of 50000, so this
should be enough for most usages.

Some documentation is lacking, but needed, especially because the main
principle is far from being obvious, as it consists in storing half-bits
and have "cover nodes" set at the lowest half of a bit and split nodes
at the highest half.