EXP: change ebtree to support relative pointers
Pointers to other nodes are now signed ints instead of pointers. The
real offset is &ptr+ptr+2, and NULL is stored as 0. The reason for "2"
is that the left branch pointer to the next branch might point to
itself if the branch pointer is the same node's leaf. That way we
can differentiate between a real branch pointer and NULL.

Note that if we used to represent EB_LEAF as 1, then we would not
need this hack of "2" above, because a pointer to the same node's
leaf would always be 1. But this would complicate eb_walk_down.

Note: we have not yet optimized away the gettag() calls to work
on the integers instead of converting the pointer.

First tests show that the code is substantially slower than the
native one, especially the deletion. The lookup is almost unaffected,
the insertion takes about 80% more time and deletion is 3 times as
long as before (but it was the fastest too and still is).