MAJOR: replace inter-node pointer with links
A "link" is a pointer from one node to another node. At the moment,
links are still regular pointers, but the purpose is to be able to
use links as relative pointers as well. In order to do this, we'll
need not only the link's value, but also its storage address so that
it serves as a basic pointer.

We don't dereference pointers directly anymore, instead we pass through
ebx_getroot() to get a troot_t* from a link, and we set them using
ebx_setlink(). The code produced is still the same though.