EXP: make tree traversal a lot faster by removing useless NULL checks
The tree is guaranteed to have a NULL only at two places :
  - root->b[EB_LEFT] when the tree is empty
  - root->b[EB_RGHT] used to detect the root

There's no need to perform the expensive NULL check over and over when
going down. So we now have two sets of functions used to convert the
pointers. The safe ones are used only when :
  - checking a tree root before going down
  - checking that we reached the tree's root when going up ; this last
    one may later go away.

The speed is substantially faster, the tree insertion now takes 302 cycles
in 64 bits trees vs 257 in native mode for the same data. Lookup is now
almost equivalent (232 vs 222). Deletion is only 50% slower (30 vs 20).

The next changes should concern cleaning up some useless conversions that
are made just to perform a check. An eb_clrtag() variant should be made
for relative pointers too.