MEDIUM: add a new function ebx_link_is_null() to check for null pointers
This function more efficiently checks for a pointer's validity because
it directly checks the relative pointer without converting it. We must
be careful when performing this check because NULL can appear at two
places :

   root->b[EB_LEFT] where it is exactly NULL
   root->b[EB_RGHT] where it is NULL optionally with the unique flag

Thus values 0 & 1 are NULL, other ones are not. The lookup time in
relative addressing mode went back to the native one and the deletion
time is only 26 cycles vs 20. Insertion is still at 300 cycles vs 257.