MEDIUM: reorder fields in the struct for relative adressing so that we never have a NULL pointer
By having the node pointer first in the struct, we cannot get a zero
offset in normal conditions, since a node pointer may never point to
the same node.

This saves us from using bit 1 to indicate the NULL, and improves
performance : insert of 1000000 values went down from 345 to 317 ms,
and lookup went down from 318 to 303 ms.

The struct format for absolute adressing was left unchanged because
when branches are displaced below, the code produced by gcc is much
less efficient due to the extra offset that it does not always manage
to optimize.