MEDIUM: improve NULL detection at top of trees
We don't need to convert the empty entry to a pointer as we can
first detect that the first branch relative pointer is 0 then
stop on this. The code gets slightly smaller and faster this way
because gcc duplicates less code around the if.

The code in relative addressing mode is now closer to what it was
in absolute mode.