TESTS: add testfunc.c to test the various functions
Sometimes there's a desire to test certain functions versus other
ones, but at the moment there is nothing convenient for this. This
new test file makes it easier to specify at build time the node
type, the insert function, the lookup function and the delete
function to be used. This allows one to compare inline functions
with non-inlined ones for example. The time is measured in CPU
cycles on x86 only. It's returned by node, and by log(node) so
that it's easier to see when the time per level increases, which
generally indicates cache misses.