REORG: move all legacy ebtree code to eb/
This is the EBtree Absolute pointer implementation ("eba"), for now,
which is also the original one. The directory will also host the
EBtree Relative Large/Medium/Small pointer implementations ("ebl",
"ebm", "ebs") in the future, so let's call this "eb" to pack
everything together.

The file compiler.h was moved to common/, and test code was moved to
tests/.