REORG: rename ebmb_node to ebxmb_node and provide compatibility layer
The ebmbtree node is now ebxmb_node, and ebtree.h forces it to be mapped
to ebmb_node so that the types are the same in files including it. Note
that ebimtree, ebsttree, ebistree, ebpttree were adjusted as well to
include ebmbtree.h in order to get the new definition.