REORG: rename ebt_root_t to ebx_troot_t and provide compatibility layer
The ebtree branch pointer is now ebx_troot_t, and ebtree.h forces it to be
mapped to eb_troot_t so that the types are the same in files including it.
Note that this type was not exported outside of ebtree, and as it is not
convenient it will later disappear.