REORG: massive change to allow multiple format inclusion
Including multiple tree node types inside the same C file proved difficult
because the defines are split across multiple include files for a same type.
The consequence of this is that the same include file may be loaded multiple
times and is needed all these times, such as in the example below :

       #include <eba32tree.h>
       #include <ebl32tree.h>
       #include <eba64tree.h>

The second include will not reinclude ebatree.h (already included) and will
not redefine ebx_ to eba_ thus will keep ebl_ instead. One solution could be
to have each .h file keep track of the various pointer sizes it was called
with, but this is quite ugly and will become tricky to maintain over the
long term.

Another solution consists in defining all symbols of a given type at once
using a single include file. In the end that's much cleaner and simpler to
maintain. And we can do this because it's a new set of files. Thus now an
application only has to include eb{a,s,m,l}tree.h.

However we keep the compatibility with the legacy eb* files for existing
applications. Each eb* file loads ebatree.h so in practice any of them will
defined all symbols at once.

All the type-independant functions (eg: cmp_bits()) were moved to a new
file : ebcommon.h.