LICENSE: relicense to MIT/X11 instead of GPL
X11 is more permissive than GPL and explicitly allows modifications to
be shipped without having to find a place to host minor patches. This
concern was reported twice already about ebtree v6, so better address
it while it's still possible. This change also allows the code to be
copied as-is into other software covered by various licenses including
GPL and BSD.

Note to all users and/or contributors : this change is made to make
your packaging work easier. If you do find a bug or add features that
can be useful to other people, please do contribute them, they're always
welcome, and this will save you from a painful rebasing work or from API
changes later.