CLEANUP: cleanup the fls/flsnz API mess
The fls/flsnz API is not really clean because the naming is inconsistent
between functions, some explicitly have "nz" in their names while others
not, and some have the argument size and the 32-bit does not, resulting
in fls_auto() being in fact what flsnz() should be expected to be.

This change renames functions so that they all hold "nz" when they don't
accept a zero, and all specify their argument sizes, except for flsnz which
supports them all.