MEDIUM: cb: add basic compact binary tree support
The cbaa_* functions now seem to work correctly, though it's still in
development mode. Delete is not implemented yet, only insert() and
lookup{,_ge,_le} are implemented. A dump function is provided as well
and it should be generalized to other trees.

For now only this version of the compact trees supporting absolute
pointers and using the node's address as the key was implemented. Since
we're performing operations on the branch pointers themselves, we'll
likely have to use different code for different pointer types, unless
we find an elegant way to cast this.