EXAMPLES: add a simple lru cache implementation based on eb64
This one provides functions to retrieve, touch and reserve a 64-bit
key and its accompanying data from an LRU cache, with a validity
domain and revision. It's supposed to be simple enough to integrate
into projects. A quick test on a core i5-3320M shows the following
numbers for various hit rates :

   cache size   100% misses  100% hits   90% hits
      1024          10M/s       12M/s        -
      4096           8M/s         -          -
     58982            -           -       2.5M/s  (with 64k dataset)
     65536            -          3M/s        -

This code puts a lot of stress on the eb64 implementation, and eb32 is
not much faster (about 6%). That's a good use case to try to optimize
them.