MINOR: add a new ebx_is_root() function
This one is used to detect that we reached the root. We used to use the
ebx_link_is_null() function for this but that prevents us from easily
adding extra information in the root node. So in order not to conflate
real NULL detection (left branch) and ROOT detection (based on the
right branch pointer) we now have a separate function. For now it still
is exactly the same as ebx_link_is_null() (except that it automatically
dereferences the right branch).