MINOR: eb: add reentrant code for absolute trees
This patch introduces the "ebar" variants for the "eba" functions and
types. For now the types and functions are exactly the same except that
the top of the tree is detected by the presence of bit 1 in the right
branch. It would be perfectly valid to reuse the same types but the risk
of silently hiding application bugs remains high, thus it's better to
have different types.

No locking is applied yet, but the main code can now be modified to
insert such locking, which will be a NOP for eba and real locks for
ebar.