MEDIUM: change the ebx_delete() API to take a root
This one will be required to support reentrant code. __ebx_delete() now
becomes the native non-reentrant function, and ebx_delete() is an inline
pointing to it and taking a pointer to the root.

All eb_delete and eb{32,64,pt,mb}_delete functions were mapped to the
non-reentrant variant to maintain API compatibility. It's worth noting
we don't have any inlined function anymore here.

The delete functions in the examples were moved to __eba32_delete() though
the root pointer is always available.