MINOR: eb_is_dup() is for internal use only
Rename it __eb_is_dup() and ensure it's used at a few places
instead of the open-coded "node.bit < 0". There are some places
where the node's bit is copied into a temporary variable both
for this and for later use and others where it's not. It's would
be worth checking if recent compilers do a better job at one or
the other and make the whole code more uniform.