MEDIUM: always set bit 1 in branch pointers
In order to implement locking, we'll need to be able to use almost all
bits of the root's right branch. Thus we need a different way to detect
that we've reached the root. Fortunately we still have bit 1 unused in
our pointers so it can be used to mark the root. But we don't want to
break applications which expect a couple of NULL pointers to serve as
a root.

Here what we do instead is that we set bit 2 for all pointers except
the root's right branch. Thanks to this, we can now safely detect that
we're on the root by leaving its bit clear.

The code is slightly larger but keeps the same performance because all
pointers were already relative so instead of having +0 or +1 we now
have +2 or +3.

The other benefit is that it unifies ebatree and ebartree's API.