TEST: add a random access performance test
The principle is the following : we create N nodes that we store in an
indexed array, and assign each of them a unique value between 0 and N-1,
and we have an empty tree. We loop for a given time doing this :

   - pick X = random number between 0 and N-1
   - lookup value X in the tree => NODE
   - if NODE exists, take NNODE = next(NODE), and PNODE=prev(NODE), then
     delete NODE from the tree and put the node back into the array
   - if NODE doesn't exist, pick it from the array and insert it in the
     tree at position X ; take NNODE = next(NODE), PNODE = prev(NODE)
   - if NNODE, validate that NNODE->key > NODE->key
   - if PNODE, validate that PNODE->key < NODE->key

This causes random lookup/insertion/removal and also uses next() and prev().
Threads are not used for now.