WIP: make eb_troot_t a scalar, not a pointer
This is much more consistent with the various operations that are run
on it. However in the process we also changed eba_link_t to unsigned
long instead of void *, which may be visible to applications explicitly
setting NULL into node_p/leaf_p.

eba32_lookup() is still 0x6a bytes long. But the code overall is slower,
15.24M/s for stress 1000 instead of 16.42 before.