MINOR: replace ebx_gettag() with ebx_get_root_flags() to get the unique flag
There used to be a mistaked in the way the unique flag was checked. It
was set directly in the eb_link_t entry in the root, but was checked
using ebx_getroot(), meaning that it adds the root's pointer to it before
checking. In practice it has no effect since the tree is supposed to be
properly aligned, but this is wrong, will not work for locks later, and
is counter-productive.

This patch adds a new function __ebx_get_root_flags() which directly
returns the flags from a root pointer, avoiding useless intermediary
operations. The relevant functions now explicitly check for the EB_UNIQUE
flag within these flags.