MINOR: remove unused __ebx_is_null()
This one was in fact still in use but sometimes ebx_is_empty() was used
instead. It's not interesting anymore to know that a pointer is null
because the sole initial purpose of this test was to check for the
root's right branch and this test has been replaced already, so that
the only remaining places are those explicitly checking for an empty
tree. Better use the empty check instead.