MEDIUM: tag the root pointer instead of testing root's right branch
Now we use a second bit in the parent pointers to indicate whether or
not we point to the root. This ensures we can use the root's right
pointer as we want since only one parent pointer in the whole tree
will have this bit set, also making the check easier.

On difficulty remains though. It has now become mandatory to apply
a mask to the side in some operations (mainly delete) to only focus
on left vs right side when the parent could possibly be the root.
The net effect is a 1% performance drop on the stress test and a
small code size increase. The sched test however is slightly faster
(0.5%). Anyway all of these vanish once ebx32_insert() gets some
likely() in the descent loop and likely is enabled for the current
compiler, so this is not a big deal for now.