WIP: ensure the tree remains locked after eb32_lookup_ge()
We'll probably need to have a set of *_locked() functions to keep the
tree fr-locked after a successful lookup. For now we emulate this by
not releasing the lock in eb32_lookup_ge() and upgrading it in
eba32_delete().

Since we didn't modify eb32_first() instead we perform a lookup_ge(0).

Now it works. It's significantly slower than the application lock but
at least it works.