MINOR: eb32sc: implement the scope-aware functions for eb32
A new kind of tree nodes is currently being developed in ebtree v7,
consisting in storing a scope in each node indicating a visibility
mask so that certain nodes are not reported on certain lookups. The
initial goal was to make this usable with a multi-thread scheduler.

Since the ebtree v7 code is completely different from v6, this patch
instead copies the minimally required functions from eb32 and ebtree
and calls them "eb32sc_*". At the moment the scope is not implemented,
it's only passed in arguments.

(cherry picked from haproxy commit ca30839a857049acc96e26a28e7515533d9ff153)
[wt: moved files to root and dropped regprm; remerge the change from
 4b21609eb ("BUILD: ebtree: improve architecture-specific alignment")]
Signed-off-by: Willy Tarreau <w@1wt.eu>