MINOR: eb32sc: update the eb32sc parent node's scope on delete
During a delete operation, if the deleted node is above its leaf's
parent, this parent will replace the node and then go up. In this
case it is important to update the new parent's scope to reflect
the presence of other branches.

It's worth noting that in theory we should precisely recompute the
exact node value, but it seems that it's not worth it for the rare
cases there is a mismatch.

(cherry picked from haproxy commit ef8d0dcefdf00b043697df971533daf01290811a)
Signed-off-by: Willy Tarreau <w@1wt.eu>