DOC: enumerate some adjustable properties of various options
The split bit is not required but helps; the typed branches are not
required but help; the up pointers are not required but help. This
enumerates the benefits and costs brought by these options and their
suitability to various use cases. Once put down, it looks like the
main use cases fall into 3 categories:
  - generic usage: full EB
  - fast indexing: pure CB with bit for fast dup insertion
  - config data: pure CB with typed pointers to seamlessly support dups

A last, unlikely one, is the zero-overhead memory allocator which might
still make sense for microcontrollers and low memory footprint systems.