WIP: doc: explain cbtree using tagged pointers
But that's not how the code currently works, it solely uses value
comparisons. However the explained algo there should work, so it
might still be valuable if it manages to save performance at the
expense of one bit per pointer. Worth noting as well that the bit
in the pointers could be the one making the difference between the
two sides if needed.