BUILD: ebtree: improve architecture-specific alignment
Enforce proper alignment on the carrying nodes where necessary.
This has the benefit of closing holes wherever possible and to
align data to the minimally required size.

The only thing it cannot close is the 32-bit hole at the end of ebmbnode
due to the required 64-bit on certain archs but at least it guarantees
that the key correctly points to the end of the node and that there is
never a hole after it.

(cherry picked from commit 45dacd51521daf8adfdb879fc585211a57287222)