OPTIM: ebmb: make ebmb_insert_prefix() keep a copy the new node's pfx
looking at a perf profile while loading a conf with a huge map, it
appeared that there was a hot spot on the access to the new node's
prefix, which is unexpectedly being reloaded for each visited node
during the tree descent. Better keep a copy of it because with large
trees that don't fit into the L3 cache the memory bandwidth is scarce.
Doing so reduces the load time from 8.0 to 7.5 seconds.

(cherry picked from haproxy commit bf13e53964228eece9d5d532943720bd9af7915b)
[wt: addressed tiny signedness issues]
Signed-off-by: Willy Tarreau <w@1wt.eu>