WIP: cb: silence absurd -Warray-bounds warning with gcc11 in lookup
passing &key for a function argument doesn't please gcc when it comes to
using that inside container_of(), it reports non-sensical warnings like:

  cb/cbatree_u32.c:104:13: warning: array subscript 0 is outside array bounds of 'u32[1]' {aka 'unsigned int[1]'} [-Warray-bounds]
    104 |         u32 key = container_of(node, struct cba_u32, node)->key;
      |             ^~~
  cb/cbatree_u32.c:228:61: note: while referencing 'key'
    228 | struct cba_node *cba_lookup_u32(struct cba_node **root, u32 key)
      |                                                         ~~~~^~~

It's probably related to the fact that it passes the argument through a
register. Just assigning this to a local variable does the job. The code
is correctly optimized and the descent function properly uses %esi with
all the tests.